DOCD-1319 - Support Time fields, using datetime.datetime
